refactor: move itwêwina-specific to crkeng.app.preferences
